// srm_ui/src/app/main/Models/etendering/ViewModels/rfq-supplier-model.ts
export class RFQSupplierModel {
     id: string;
     rFQId: string;
     supplierID: string;
     supplierContactID: string;
     supplierGroupId: string;
     email: string;
     supplierName: string;
     invitationDate: Date;
     inviteDate: string;
     otherEmails: string;
     currencyId: string;
     isFromIFS: boolean;
     isEmergencySupplier: boolean;
}